+++ strato/core/blockapps-mpdbs/src/Blockchain/DB/CodeDB.hs
@@ -1,6 +1,8 @@
 {-# LANGUAGE ConstraintKinds #-}
+{-# LANGUAGE DeriveFunctor #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeApplications #-}
@@ -46,9 +48,26 @@ newtype CodeDB = CodeDB {unCodeDB :: DB.DB}
 instance NFData CodeDB where
   rnf (CodeDB a) = a `seq` ()
 
-type HasCodeDB m = (Keccak256 `Alters` DBCode) m
+type HasCodeDB m = (Keccak256 `A.Alters` DBCode) m
 
 type DBCode = (CodeKind, B.ByteString)
+
+newtype MemCodeDB m a = MemCodeDB { unMemCodeDB :: StateT (M.Map Keccak256 DBCode) m a }
+  deriving (Functor, Applicative, Monad, MonadIO)
+
+instance MonadTrans MemCodeDB where
+  lift = MemCodeDB . lift
+
+instance Monad m => (Keccak256 `A.Alters` DBCode) (MemCodeDB m) where
+  lookup _   = MemCodeDB . gets . M.lookup
+  insert _ k = MemCodeDB . modify' . M.insert k
+  delete _   = MemCodeDB . modify' . M.delete
+
+runMemCodeDB :: Monad m => MemCodeDB m a -> M.Map Keccak256 DBCode -> m a
+runMemCodeDB f m = evalStateT (unMemCodeDB f) m
+
+runNewMemCodeDB :: Monad m => MemCodeDB m a -> m a
+runNewMemCodeDB f = runMemCodeDB f M.empty
 
 shaToKey :: Keccak256 -> B.ByteString
 shaToKey = BL.toStrict . encode . sha2StateRoot
@@ -75,7 +94,7 @@ genericDeleteCodeDB f codeHash = do
   db <- unCodeDB <$> f
   DB.delete db def (shaToKey codeHash)
 
-instance MonadIO m => (Keccak256 `Alters` DBCode) (ReaderT CodeDB m) where
+instance MonadIO m => (Keccak256 `A.Alters` DBCode) (ReaderT CodeDB m) where
   lookup _ = genericLookupCodeDB ask
   insert _ = genericInsertCodeDB ask
   delete _ = genericDeleteCodeDB ask
@@ -105,4 +124,4 @@ codeDBPut kind code = do
   return hsh
 
 codeDBGet :: HasCodeDB m => Keccak256 -> m (Maybe DBCode)
-codeDBGet = lookup Proxy
+codeDBGet = A.lookup A.Proxy