+++ strato/core/vm-tools/src/Blockchain/MemVMContext.hs
@@ -272,6 +272,9 @@ instance (Account `A.Alters` AddressState) MemContextM where
   insert _ = putAddressState
   delete _ = deleteAddressState
 
+instance A.Selectable Account AddressState MemContextM where
+  select _ = getAddressStateMaybe
+
 instance (Maybe Word256 `A.Alters` MP.StateRoot) MemContextM where
   lookup _ chainId = do
     mBH <- gets $ view $ memDBs . currentBlock